requirements.txt
openai
gradio
python-dotenv
langchain
langchain-openai
